+++ src/config/images.ts
@@ -1,10 +1,14 @@
 //section background images
-
 import bg1 from '../assets/images/background.png';
 import bg2 from '../assets/images/bg2.png';
 import bg3 from '../assets/images/bg3.png';
 import bg4 from '../assets/images/bg4.png';
+import bg5 from '../assets/images/bg5.png';
 import noise from '../assets/images/noise.png';
+
+//site assets
+import icon from '../assets/images/icon.png';
+import icon_dark from '../assets/images/icon_dark.png';
 
 
 //google icons
@@ -18,6 +22,10 @@ import forum from '../assets/icons/forum.svg';
 import social from '../assets/icons/social.svg';
 import marketing from '../assets/icons/marketing.svg';
 import design from '../assets/icons/design.svg';
+import radio_checked from '../assets/icons/radio_checked.svg';
+import radio_unchecked from '../assets/icons/radio_unchecked.svg';
+import close from '../assets/icons/close.svg';
+import add from '../assets/icons/add.svg';
 
 //tech images
 import react from '../assets/logos/react.png';
@@ -28,5 +36,32 @@ import typescript from '../assets/logos/typescript.png';
 import vite from '../assets/logos/vite.png';
 
 export {
-   arrow, barrier, cached, check, code, mail, forum, social, marketing, design, bg1, bg2, bg3, bg4, noise, react, tailwind, typescript, vite, pwa, firebase
+   arrow,
+   barrier,
+   cached,
+   check,
+   code,
+   mail,
+   forum,
+   social,
+   marketing,
+   design,
+   bg1,
+   bg2,
+   bg3,
+   bg4,
+   noise,
+   react,
+   tailwind,
+   typescript,
+   vite,
+   pwa,
+   firebase,
+   icon,
+   icon_dark,
+   radio_checked,
+   radio_unchecked,
+   bg5,
+   close,
+   add
 }